Refactor notification system structure for sendgrid prep
