+++ notificationSystem/runTextNotifications.js
@@ -1,7 +1,4 @@
-/*
- * Send messages to all appointment owners via Twilio
- * @param {array} appointments List of appointments.
- */
+// Twilio API, Send incoming notifications as text message
 
 require('dotenv').config();
 const twilio = require('twilio');
@@ -9,21 +6,21 @@ const twilio = require('twilio');
 const accountSid = process.env.TWILIO_SID;
 const authToken = process.env.TWILIO_TOKEN;
 
-function sendNotifications(notifications) {
+function sendTextNotifications(notifications) {
   const client = new twilio(accountSid, authToken);
   notifications.forEach(notification => {
     // Create options to send the message
     const options = {
       to: notification.phoneNumber,
       from: process.env.TWILIO_NUMBER,
-      body: `${notification.postName}  ${notification.postDetails} Read more: ${notification.link}`
+      body: `${notification.postName}  ${notification.postDetails} Read more: ${
+        notification.link
+      }`
     };
 
     // Send the message!
-    // console.log(options);
-    client.messages.create(options, function (err, response) {
+    client.messages.create(options, function(err, response) {
       if (err) {
-        // Just log it for now
         console.error(err);
       } else {
         // Log the last few digits of a phone number
@@ -33,14 +30,6 @@ function sendNotifications(notifications) {
       }
     });
   });
-
-  // Don't wait on success/failure, just indicate all messages have been
-  // queued for delivery
-  // if (callback) {
-  //   callback.call();
-  // }
 }
 
-// sendNotifications();
-
-module.exports = sendNotifications;
+module.exports = sendTextNotifications;
