+++ package.json
@@ -19,6 +19,7 @@
 	},
 	"homepage": "https://github.com/Blent1050/auth-ii#readme",
 	"dependencies": {
+		"@sendgrid/mail": "^6.3.1",
 		"bcrypt": "^3.0.4",
 		"connect-session-knex": "^1.4.0",
 		"cors": "^2.8.5",
